// frontend/lib/api/generated/model/alertasAlmacenOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { ObservacionSinResolverOut } from './observacionSinResolverOut';
import type { ProximoVencerOut } from './proximoVencerOut';
import type { StockBajoOut } from './stockBajoOut';

export interface AlertasAlmacenOut {
  stock_bajo: StockBajoOut[];
  proximos_vencer: ProximoVencerOut[];
  observaciones_sin_resolver: ObservacionSinResolverOut[];
}