// frontend/lib/api/generated/model/valorizacionAlmacenOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface ValorizacionAlmacenOut {
  almacen_id: number;
  almacen_codigo: string;
  almacen_nombre: string;
  valor_valorizado_total: number;
}